requirements.txt
flask>=3.0.0,<4.0
waitress>=2.1.2

groq>=0.9.0

pdfplumber>=0.10.3
python-docx>=1.1.0
python-pptx>=0.6.23

langchain-text-splitters>=0.0.1

Werkzeug>=3.0.0
